Restructured ChessServer\Game
